# src/auth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "列出所有用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "过滤条件",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页尺寸",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"page\":1,\"page_size\":50,\"pages\":1,\"users\":[{\"id\":1,\"username\":\"username\",\"email\":\"email\",\"phone\":\"phone\",\"is_superuser\":false,\"disabled\":false,\"last_login\":null,\"created_at\":\"2021-01-08 10:57:27\",\"updated_at\":\"2021-01-26 11:31:24\"}],\"total\":1}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "put": {
                "tags": [
                    "用户"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"user\":{\"id\":1,\"username\":\"username\",\"email\":\"email\",\"phone\":\"phone\",\"is_superuser\":false,\"disabled\":false,\"last_login\":null,\"created_at\":\"2021-01-08 10:57:27\",\"updated_at\":\"2021-01-26 11:31:24\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/department": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "获得用户所在部门（一个用户只能在一个部门）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"departments\":{\"id\":2,\"name\":\"dept2\",\"is_owner\":true,\"joined_at\":\"2021-01-21 11:20:56\"},\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/groups": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "列出用户所在组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"groups\":[{\"id\":2,\"name\":\"group2\",\"is_owner\":true,\"joined_at\":\"2021-01-21 11:20:56\"}],\"message\":\"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/products": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "列出用户所在产品线",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"products\":[{\"id\":3,\"name\":\"SCM\",\"is_owner\":false,\"joined_at\":\"0001-01-01 00:00:00\",\"alias\":\"scm\",\"disable\":false}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/roles": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "列出用户所在角色",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"Success\",\"roles\":[{\"id\":1,\"name\":\"auth_admin\"},{\"id\":2,\"name\":\"tester\"}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "database.User": {
            "type": "object",
            "required": [
                "email",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    }
}